backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-multipart
pydantic
opencv-python-headless
ultralytics
python-dotenv
cloudinary
websockets
